Add window/progress notification
