+++ haskell-lsp-types/haskell-lsp-types.cabal
@@ -39,6 +39,7 @@ library
                      , Language.Haskell.LSP.Types.Symbol
                      , Language.Haskell.LSP.Types.TextDocument
                      , Language.Haskell.LSP.Types.Uri
+                     , Language.Haskell.LSP.Types.Window
                      , Language.Haskell.LSP.Types.WorkspaceEdit
                      , Language.Haskell.LSP.Types.WorkspaceFolders
  -- other-extensions:
