+++ haskell-lsp-types/src/Language/Haskell/LSP/Types.hs
@@ -14,6 +14,7 @@ module Language.Haskell.LSP.Types
   , module Language.Haskell.LSP.Types.Symbol
   , module Language.Haskell.LSP.Types.TextDocument
   , module Language.Haskell.LSP.Types.Uri
+  , module Language.Haskell.LSP.Types.Window
   , module Language.Haskell.LSP.Types.WorkspaceEdit
   , module Language.Haskell.LSP.Types.WorkspaceFolders
   )
@@ -34,5 +35,6 @@ import           Language.Haskell.LSP.Types.Message
 import           Language.Haskell.LSP.Types.Symbol
 import           Language.Haskell.LSP.Types.TextDocument
 import           Language.Haskell.LSP.Types.Uri
+import           Language.Haskell.LSP.Types.Window
 import           Language.Haskell.LSP.Types.WorkspaceEdit
 import           Language.Haskell.LSP.Types.WorkspaceFolders
