+++ haskell-lsp-types/src/Language/Haskell/LSP/Types/DataTypesJSON.hs
@@ -813,193 +813,6 @@ type ExitNotification = NotificationMessage ClientMethod (Maybe ExitParams)
 
 -- ---------------------------------------------------------------------
 {-
-ShowMessage Notification
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#showmessage-notification
-
-The show message notification is sent from a server to a client to ask the
-client to display a particular message in the user interface.
-
-Notification:
-
-    method: 'window/showMessage'
-    params: ShowMessageParams defined as follows:
-
-interface ShowMessageParams {
-    /**
-     * The message type. See {@link MessageType}.
-     */
-    type: number;
-
-    /**
-     * The actual message.
-     */
-    message: string;
-}
-
-Where the type is defined as follows:
-
-enum MessageType {
-    /**
-     * An error message.
-     */
-    Error = 1,
-    /**
-     * A warning message.
-     */
-    Warning = 2,
-    /**
-     * An information message.
-     */
-    Info = 3,
-    /**
-     * A log message.
-     */
-    Log = 4
-}
--}
-data MessageType = MtError   -- ^ Error = 1,
-                 | MtWarning -- ^ Warning = 2,
-                 | MtInfo    -- ^ Info = 3,
-                 | MtLog     -- ^ Log = 4
-        deriving (Eq,Ord,Show,Read,Enum)
-
-instance A.ToJSON MessageType where
-  toJSON MtError   = A.Number 1
-  toJSON MtWarning = A.Number 2
-  toJSON MtInfo    = A.Number 3
-  toJSON MtLog     = A.Number 4
-
-instance A.FromJSON MessageType where
-  parseJSON (A.Number 1) = pure MtError
-  parseJSON (A.Number 2) = pure MtWarning
-  parseJSON (A.Number 3) = pure MtInfo
-  parseJSON (A.Number 4) = pure MtLog
-  parseJSON _            = mempty
-
--- ---------------------------------------
-
-
-data ShowMessageParams =
-  ShowMessageParams {
-    _xtype   :: MessageType
-  , _message :: Text
-  } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions{ fieldLabelModifier = customModifier } ''ShowMessageParams
-
-type ShowMessageNotification = NotificationMessage ServerMethod ShowMessageParams
-
--- ---------------------------------------------------------------------
-{-
-ShowMessage Request
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#showmessage-request
-
-    New: The show message request is sent from a server to a client to ask the
-    client to display a particular message in the user interface. In addition to
-    the show message notification the request allows to pass actions and to wait
-    for an answer from the client.
-
-Request:
-
-    method: 'window/showMessageRequest'
-    params: ShowMessageRequestParams defined as follows:
-
-Response:
-
-    result: the selected MessageActionItem
-    error: code and message set in case an exception happens during showing a message.
-
-interface ShowMessageRequestParams {
-    /**
-     * The message type. See {@link MessageType}
-     */
-    type: number;
-
-    /**
-     * The actual message
-     */
-    message: string;
-
-    /**
-     * The message action items to present.
-     */
-    actions?: MessageActionItem[];
-}
-
-Where the MessageActionItem is defined as follows:
-
-interface MessageActionItem {
-    /**
-     * A short title like 'Retry', 'Open Log' etc.
-     */
-    title: string;
-}
--}
-
-data MessageActionItem =
-  MessageActionItem
-    { _title :: Text
-    } deriving (Show,Read,Eq)
-
-deriveJSON lspOptions ''MessageActionItem
-
-
-data ShowMessageRequestParams =
-  ShowMessageRequestParams
-    { _xtype   :: MessageType
-    , _message :: Text
-    , _actions :: Maybe [MessageActionItem]
-    } deriving (Show,Read,Eq)
-
-deriveJSON lspOptions{ fieldLabelModifier = customModifier } ''ShowMessageRequestParams
-
-type ShowMessageRequest = RequestMessage ServerMethod ShowMessageRequestParams Text
-type ShowMessageResponse = ResponseMessage Text
-
--- ---------------------------------------------------------------------
-{-
-LogMessage Notification
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#logmessage-notification
-
-The log message notification is sent from the server to the client to ask the
-client to log a particular message.
-
-Notification:
-
-    method: 'window/logMessage'
-    params: LogMessageParams defined as follows:
-
-interface LogMessageParams {
-    /**
-     * The message type. See {@link MessageType}
-     */
-    type: number;
-
-    /**
-     * The actual message
-     */
-    message: string;
-}
-
-Where type is defined as above.
--}
-
-data LogMessageParams =
-  LogMessageParams {
-    _xtype   :: MessageType
-  , _message :: Text
-  } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions{ fieldLabelModifier = customModifier } ''LogMessageParams
-
-
-type LogMessageNotification = NotificationMessage ServerMethod LogMessageParams
-
--- ---------------------------------------------------------------------
-{-
 Telemetry Notification
 
 https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#telemetry-notification
